Refactor package naming
