+++ .idea/gradle.xml
@@ -12,7 +12,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/panel" />
+            <option value="$PROJECT_DIR$/panellayout" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
