+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion 29
 
     defaultConfig {
-        applicationId "com.wayfair.panellayout"
+        applicationId "com.wayfair.panellayout.sample"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
@@ -30,7 +30,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
-    implementation project(":panel")
+    implementation project(':panellayout')
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
