+++ panellayout/src/main/kotlin/com/wayfair/panellayout/PanelPosition.kt
@@ -1,4 +1,4 @@
-package com.wayfair.waystation.panel
+package com.wayfair.panellayout
 
 enum class PanelPosition {
     LEFT_EDGE, RIGHT_EDGE, TOP_EDGE, BOTTOM_EDGE, NO_EDGE
