+++ panellayout/src/main/res/layout/panel_view_default_snap_bottom.xml
