+++ panellayout/src/main/res/values/panel_attrs.xml
