+++ panellayout/src/main/res/values/panel_colors.xml
