+++ panellayout/src/main/res/values/panel_dimens.xml
