+++ panellayout/src/main/res/values/panel_integers.xml
